Remove Blend, fix color spaces, add lighten, darken and mix node
