+++ .changeset/weak-pugs-sneeze.md
@@ -1,0 +1,5 @@
+---
+"@tokens-studio/graph-engine": major
+---
+
+Remove Blend Node and split it up into Lighten, Darken, Mix nodes
