--- packages/graph-engine/src/nodes/color/lib/mix.ts
@@ -1,13 +1,0 @@
-import { ColorSpaceTypes } from "./types.js";
-import Color from "colorjs.io";
-
-export function mix(
-  color: Color,
-  colorSpace: ColorSpaceTypes,
-  amount: number,
-  mixColor: Color
-) {
-  const mixValue = Math.max(0, Math.min(1, Number(amount)));
-
-  return Color.mix(color, mixColor, mixValue);
-}
