+++ packages/graph-engine/src/nodes/color/lib/types.ts
@@ -1,42 +1,39 @@
-import { ColorModifierTypes } from "@tokens-studio/types";
+export const colorSpaces = [
+  "a98rgb",
+  "acescc",
+  "acescg",
+  // "cam16",
+  "hct",
+  "hpluv",
+  "hsl",
+  "hsluv",
+  "hsv",
+  "hwb",
+  "ictcp",
+  "jzazbz",
+  "jzczhz",
+  "lab-d65",
+  "lab",
+  "lch",
+  "lchuv",
+  "luv",
+  // "okhsl",
+  // "okhsv",
+  "oklab",
+  "oklch",
+  "p3-linear",
+  "p3",
+  "prophoto-linear",
+  "prophoto",
+  "rec2020-linear",
+  "rec2020",
+  // "rec2100-hlg",
+  // "rec2100-pq",
+  "srgb-linear",
+  "srgb",
+  "xyz-abs-d65",
+  "xyz-d50",
+  "xyz-d65"
+] as const;
 
-export enum ColorSpaceTypes {
-  OKLAB = "oklab",
-  OKLCH = "oklch",
-  LCH = "lch",
-  SRGB = "srgb",
-  P3 = "p3",
-  HSL = "hsl",
-}
-
-interface Modifier<T extends string, V> {
-  type: T;
-  value: V;
-}
-interface ColorGenericModifier<T extends ColorModifierTypes, V>
-  extends Modifier<T, V> {
-  space: ColorSpaceTypes;
-}
-
-export type LightenModifier = ColorGenericModifier<
-  ColorModifierTypes.LIGHTEN,
-  string
->;
-export type DarkenModifier = ColorGenericModifier<
-  ColorModifierTypes.DARKEN,
-  string
->;
-export interface MixModifier
-  extends ColorGenericModifier<ColorModifierTypes.MIX, string> {
-  color: string;
-}
-export type AlphaModifier = ColorGenericModifier<
-  ColorModifierTypes.ALPHA,
-  string
->;
-
-export type ColorModifier =
-  | LightenModifier
-  | DarkenModifier
-  | MixModifier
-  | AlphaModifier;
+export type ColorSpace = typeof colorSpaces[number];
