+++ packages/graph-engine/src/nodes/color/lighten.ts
@@ -1,20 +1,15 @@
-import {
-  ColorModifier,
-  ColorSpaceTypes,
-} from "@tokens-studio/types";
+import { Black, toColor, toColorObject } from "./lib/utils.js";
 import {
   ColorSchema,
   NumberSchema,
   StringSchema,
 } from "../../schemas/index.js";
+import { ColorSpace, colorSpaces } from "./lib/types.js"
 import { Color as ColorType } from "../../types.js";
 import { INodeDefinition, ToInput, ToOutput } from "../../index.js";
 import { Node } from "../../programmatic/node.js";
-import { convertModifiedColorToHex } from "./lib/modifyColor.js";
-import { toColor, toColorObject } from "./lib/utils.js";
-
+import Color from "colorjs.io";
 export { ColorModifierTypes } from "@tokens-studio/types";
-
 
 export default class NodeDefinition extends Node {
   static title = "Lighten Color";
@@ -24,7 +19,7 @@ export default class NodeDefinition extends Node {
   declare inputs: ToInput<{
     color: ColorType;
     value: number;
-    space: ColorSpaceTypes;
+    space: ColorSpace;
   }>;
 
   declare outputs: ToOutput<{
@@ -36,7 +31,7 @@ export default class NodeDefinition extends Node {
     this.addInput("color", {
       type: { 
         ...ColorSchema,
-        default: "#000000"
+        default: Black
       },
     });
     this.addInput("value", {
@@ -46,14 +41,6 @@ export default class NodeDefinition extends Node {
         description: "Value to apply to the modifier",
       },
     });
-    this.addInput("space", {
-      type: {
-        ...StringSchema,
-        default: "srgb",
-        enum: Object.keys(ColorSpaceTypes),
-        description: "The color space we are operating in",
-      },
-    });
 
     this.addOutput("value", {
       type: ColorSchema,
@@ -61,18 +48,14 @@ export default class NodeDefinition extends Node {
   }
 
   execute(): void | Promise<void> {
-    const { space, value, color } = this.getAllInputs();
+    const { value, color } = this.getAllInputs();
+    
+    const sourceColor = toColor(color);
+    const lightness = sourceColor.oklch.l;
 
-    const col = toColor(color)
-
-    const converted = convertModifiedColorToHex(col, {
-      type: "lighten",
-      space,
-      value,
-    } as ColorModifier);
-
-    const final = toColorObject(converted);
-
+    const newLightness = lightness + (1 - lightness) * value;
+    sourceColor.oklch.l = newLightness;
+    const final = toColorObject(sourceColor);
 
     this.setOutput("value", final);
   }
